+++ Headers/DebugServer2/Utils/SwapEndian.h
@@ -13,15 +13,15 @@
 
 namespace ds2 {
 
-static inline uint16_t Swap16(uint16_t x) { return (x >> 8) | (x << 8); }
+static inline uint16_t Swap16(uint16_t x) { return static_cast<uint16_t>((x >> 8) | (x << 8)); }
 
 static inline uint32_t Swap32(uint32_t x) {
-  return Swap16(x >> 16) | (Swap16(x & 0xffff) << 16);
+  return static_cast<uint32_t>(Swap16(static_cast<uint16_t>(x >> 16)) | (Swap16(x & 0xffff) << 16));
 }
 
 static inline uint64_t Swap64(uint64_t x) {
-  return Swap32(x >> 32) |
-         (static_cast<uint64_t>(Swap32(x & 0xffffffff)) << 32);
+  return Swap32(static_cast<uint32_t>(x >> 32)) |
+         (static_cast<uint64_t>(Swap32(static_cast<uint32_t>(x & 0xffffffff))) << 32);
 }
 }
